+++ src/main/java/com/sparta/blogproject/like/dto/MsgResponseDto.java
@@ -9,7 +9,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-
 public class MsgResponseDto {
     //필드
     private String msg;
