--- src/main/java/com/sparta/blogproject/like/repository/CommentRepository.java
@@ -1,10 +1,0 @@
-package com.sparta.blogproject.like.repository;
-
-import com.sparta.blogproject.like.entity.Comment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface CommentRepository extends JpaRepository<Comment, Long> {
-    Optional<Comment> findByIdAndUserId(Long commentId, Long UserId);
-}
