+++ src/main/java/com/sparta/blogproject/like/service/LikeService.java
@@ -1,24 +1,27 @@
 package com.sparta.blogproject.like.service;
 
-import com.sparta.blogproject.like.entity.Comment;
-import com.sparta.blogproject.like.repository.CommentLikeRepository;
-import com.sparta.blogproject.like.repository.CommentRepository;
 import com.sparta.blogproject.like.dto.MsgResponseDto;
+import com.sparta.blogproject.like.entity.Comment;
 import com.sparta.blogproject.like.entity.CommentLike;
+import com.sparta.blogproject.like.entity.PostLike;
+import com.sparta.blogproject.like.repository.CommentLikeRepository;
+//import com.sparta.blogproject.like.repository.CommentRepository; TODO : 추후 수정 예정
+import com.sparta.blogproject.like.repository.PostLikeRepository;
+import com.sparta.blogproject.post.entity.Post;
 import com.sparta.blogproject.post.repository.PostRepository;
+import com.sparta.blogproject.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
-import com.sparta.blogproject.user.entity.User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-
 @Service
 @RequiredArgsConstructor
-public class CommentLikeService {
-    private final CommentRepository commentRepository;
+public class LikeService {
+//    private final CommentRepository commentRepository;    TODO : 추후 수정 예정
     private final PostRepository postRepository;
     private final CommentLikeRepository commentLikeRepository;
+    private final PostLikeRepository postLikeRepository;
 
     //댓글 좋아요
     @Transactional
@@ -38,6 +41,28 @@ public class CommentLikeService {
             //DB 에 해당 user 가 '댓글 좋아요' 누른 적이 있다면, '댓글 좋아요' 를 제거하기
         }else {
             commentLikeRepository.deleteByCommentIdAndUserId(comment.getId(), Long.valueOf(user.getUsername()));
+            return new MsgResponseDto("좋아요 취소", HttpStatus.OK.value());
+        }
+    }
+
+    // 게시글 좋아요 확인
+    @Transactional(readOnly = true)
+    public boolean checkPostLike(Long postId, User user) {
+        return postLikeRepository.existsByPostIdAndUserId(postId, Long.valueOf(user.getUsername()));
+    }
+
+    // 게시글 좋아요 생성 및 삭제
+    @Transactional
+    public MsgResponseDto savePostLike(Long postId, User user) {
+        Post post = postRepository.findById(postId).orElseThrow(
+                () -> new IllegalArgumentException("찾는 게시글이 없습니다.")
+        );
+
+        if (!checkPostLike(postId, user)) {
+            postLikeRepository.saveAndFlush(new PostLike(post, user));
+            return new MsgResponseDto("좋아요 완료", HttpStatus.OK.value());
+        } else {
+            postLikeRepository.deleteByPostIdAndUserId(postId, Long.valueOf(user.getUsername()));
             return new MsgResponseDto("좋아요 취소", HttpStatus.OK.value());
         }
     }
